flask_docker_2/webapp/requirements.txt
Flask==2.1.1
grpcio==1.51.3
grpcio-tools==1.51.3
requests==2.28.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
zipp==3.8.0
